Integrated the new summary style into the TextDocumentService code
